+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Version 1-SNAPSHOT
 ADDED:
 
 - new CHANGELOG.md file created
+- new converter classes for transform from xml string to object and back
 
 Notable links:
 [keep a changelog](http://keepachangelog.com/en/1.0.0/) Don’t let your friends dump git logs into changelogs